refactor: add `as_conversions` clippy correctness lint

+++ crates/ironrdp-web/src/session.rs
@@ -69,7 +69,7 @@ struct SessionBuilderInner {
 
     use_display_control: bool,
     enable_credssp: bool,
-    outbound_message_size_limit: Option<u32>,
+    outbound_message_size_limit: Option<usize>,
 }
 
 impl Default for SessionBuilderInner {
@@ -216,8 +216,8 @@ impl iron_remote_desktop::SessionBuilder for SessionBuilder {
             |enable_credssp: bool| { self.0.borrow_mut().enable_credssp = enable_credssp };
             |outbound_message_size_limit: f64| {
                 let limit = if outbound_message_size_limit >= 0.0 && outbound_message_size_limit <= f64::from(u32::MAX) {
-                    #[expect(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-                    { outbound_message_size_limit as u32 }
+                    #[expect(clippy::as_conversions, clippy::cast_possible_truncation, clippy::cast_sign_loss)]
+                    { outbound_message_size_limit as usize }
                 } else {
                     warn!(outbound_message_size_limit, "Invalid outbound message size limit; fallback to unlimited");
                     0 // Fallback to no limit for invalid values.
@@ -899,20 +899,20 @@ fn build_config(
 async fn writer_task(
     rx: mpsc::UnboundedReceiver<Vec<u8>>,
     rdp_writer: WriteHalf<WebSocket>,
-    outbound_limit: Option<u32>,
+    outbound_limit: Option<usize>,
 ) {
     debug!("writer task started");
 
     async fn inner(
         mut rx: mpsc::UnboundedReceiver<Vec<u8>>,
         mut rdp_writer: WriteHalf<WebSocket>,
-        outbound_limit: Option<u32>,
+        outbound_limit: Option<usize>,
     ) -> anyhow::Result<()> {
         while let Some(frame) = rx.next().await {
             match outbound_limit {
-                Some(max_size) if frame.len() > max_size as usize => {
+                Some(max_size) if frame.len() > max_size => {
                     // Send in chunks.
-                    for chunk in frame.chunks(max_size as usize) {
+                    for chunk in frame.chunks(max_size) {
                         rdp_writer.write_all(chunk).await.context("couldn't write chunk")?;
                         rdp_writer.flush().await.context("couldn't flush chunk")?;
                     }
@@ -1153,6 +1153,7 @@ where
     }
 }
 
+#[expect(clippy::as_conversions)]
 #[expect(clippy::cast_sign_loss)]
 #[expect(clippy::cast_possible_truncation)]
 fn f64_to_u16_saturating_cast(value: f64) -> u16 {
